refactor(ui-react): extract tool preview utilities
